Moved src/HostRepository to src/types/HostRepository
